730 - updated with Ivanos changes and refactored
