+++ src/main/scala/tech/cryptonomic/conseil/Lorre.scala
@@ -7,6 +7,7 @@ import akka.stream.ActorMaterializer
 import mouse.any._
 import com.typesafe.scalalogging.LazyLogging
 import org.slf4j.LoggerFactory
+import slick.lifted.{AbstractTable, TableQuery}
 import tech.cryptonomic.conseil.tezos.{
   ApiOperations,
   DatabaseConversions,
@@ -23,6 +24,7 @@ import tech.cryptonomic.conseil.tezos.TezosTypes._
 import tech.cryptonomic.conseil.io.MainOutputs.LorreOutput
 import tech.cryptonomic.conseil.util.{ConfigUtil, DatabaseUtil}
 import tech.cryptonomic.conseil.config._
+import tech.cryptonomic.conseil.tezos.TezosDatabaseOperations.insertWhenEmpty
 import tech.cryptonomic.conseil.tezos.TezosNodeOperator.{FetchRights, LazyPages}
 
 import scala.concurrent.duration._
@@ -81,19 +83,37 @@ object Lorre extends App with TezosErrors with LazyLogging with LorreAppConfig w
   )
 
   /** Inits registered tokens at startup */
-  import kantan.csv.generic._
-
-  val tokenRows: List[Tables.RegisteredTokensRow] =
-    ConfigUtil.Csv.readTableRowsFromCsv(Tables.RegisteredTokens, tezosConf.network, separator = '|')
-
-  db.run(TezosDb.writeRegisteredTokensRowsIfEmpty(tokenRows)) andThen {
-      case Success(_) => logger.info(s"Written ${tokenRows.size} registered token rows")
-      case Failure(e) => logger.error("Could not fill registered_tokens table", e)
-    }
+//  import kantan.csv.generic._
+//
+//  val tokenRows: List[Tables.RegisteredTokensRow] =
+//    ConfigUtil.Csv.readTableRowsFromCsv(Tables.RegisteredTokens, tezosConf.network, separator = '|')
+//
+//  db.run(TezosDb.insertWhenEmpty(Tables.RegisteredTokens, tokenRows)) andThen {
+//      case Success(_) => logger.info(s"Written ${tokenRows.size} registered token rows")
+//      case Failure(e) => logger.error("Could not fill registered_tokens table", e)
+//    }
   /** Inits tables with values from CSV files */
   import kantan.csv.generic._
-  TezosDb.initTableFromCsv(Tables.RegisteredTokens, tezosConf.network, separator = '|')
-  TezosDb.initTableFromCsv(Tables.KnownAddresses, tezosConf.network)
+  initTableFromCsv(Tables.RegisteredTokens, tezosConf.network, separator = '|')
+  initTableFromCsv(Tables.KnownAddresses, tezosConf.network)
+
+  import shapeless._
+  import shapeless.ops.hlist._
+
+  import kantan.csv._
+
+  /** Reads and inserts CSV file to the database for the given table */
+  def initTableFromCsv[A <: AbstractTable[_], H <: HList](table: TableQuery[A], network: String, separator: Char = ',')(
+      implicit hd: HeaderDecoder[A#TableElementType],
+      g: Generic.Aux[A#TableElementType, H],
+      m: Mapper.Aux[tech.cryptonomic.conseil.util.ConfigUtil.Csv.Trimmer.type, H, H]
+  ): Future[Option[ProposalSupporters]] = {
+    val rows = ConfigUtil.Csv.readTableRowsFromCsv(table, network, separator)
+    db.run(insertWhenEmpty(table, rows)) andThen {
+      case Success(_) => logger.info(s"Written ${rows.size} ${table.baseTableRow.tableName} rows")
+      case Failure(e) => logger.error(s"Could not fill ${table.baseTableRow.tableName} table", e)
+    }
+  }
 
   /** Schedules method for fetching baking rights */
   system.scheduler.schedule(lorreConf.blockRightsFetching.initDelay, lorreConf.blockRightsFetching.interval)(
